requirements.txt
python-gsmmodem-new==0.13.0
fastapi==0.78.0
fastapi-key-auth==0.7.5